u64 is now the architecture size. memory is [u8;MEM_SIZE]. Arithmetic Instructions need fixing. Doc needs changes. Added ldr and str instructions for all integer sizes (byte (u8), halfword (u16), word (u32), doubleword (u64), quadword (u128). ldr/str are synonyms too ldrd/strd for now

+++ procasm/src/instruction/memory_location.rs
@@ -4,7 +4,7 @@ use crate::instruction::{Instruction, operand::Operand};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum MemoryLocation {
-    Labeled(usize),
+    Labeled(u64),
     Offset { base: Register, offset: Operand },
 }
 
@@ -14,7 +14,7 @@ impl MemoryLocation {
     pub fn resolve<const MEM_SIZE: usize, Insts, Bytes>(
         self,
         processor: &Processor<MEM_SIZE, Instruction, Insts, Bytes>,
-    ) -> usize {
+    ) -> u64 {
         match self {
             Self::Labeled(addr) => addr,
             Self::Offset { base, offset } => processor.registers.get_reg(base) + offset.resolve(processor),
